[refactor] 패키지 정리(RacingCar)
