+++ src/main/java/racingcar/domain/MotorRacing.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
 
+import racingcar.domain.core.RacingCar;
 import racingcar.ui.view.MotorRacingDisplayResults;
 import racingcar.ui.view.MotorRacingDisplayRoundResult;
