+++ src/main/java/racingcar/domain/core/IncreaseMileageRacingCar.java
@@ -1,4 +1,7 @@
-package racingcar.domain;
+package racingcar.domain.core;
+
+import racingcar.domain.MotorRacingRule;
+import racingcar.domain.RacingResult;
 
 import static racingcar.domain.RacingResult.defaultResult;
