+++ src/test/java/racingcar/domain/core/RacingCarTest.java
@@ -1,9 +1,12 @@
-package racingcar.domain;
+package racingcar.domain.core;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import racingcar.domain.MotorRacingRule;
+import racingcar.domain.RacingResult;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
